requirements.txt
# --- Core API ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# --- Base de Datos SQL (Postgres) ---
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# --- Base de Datos NoSQL (Mongo) ---
pymongo>=4.6.1
motor>=3.3.0

# --- Seguridad y Autenticación ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# --- AI ---
google-genai>=1.0.0

# --- Frontend ---
aiofiles>=23.0.0